+++ cp_generator/generator/src/main/java/eu/paasage/upperware/profiler/generator/service/camel/impl/NodeCandidateServiceImpl.java
@@ -1,7 +1,10 @@
 package eu.paasage.upperware.profiler.generator.service.camel.impl;
 
-import eu.melodic.cloudiator.client.model.*;
 import eu.paasage.upperware.profiler.generator.service.camel.NodeCandidatesService;
+import io.github.cloudiator.rest.model.Hardware;
+import io.github.cloudiator.rest.model.Image;
+import io.github.cloudiator.rest.model.NodeCandidate;
+import io.github.cloudiator.rest.model.OperatingSystem;
 import org.apache.commons.lang3.tuple.ImmutablePair;
 import org.apache.commons.lang3.tuple.Pair;
 import org.springframework.stereotype.Service;
@@ -34,10 +37,10 @@ public class NodeCandidateServiceImpl implements NodeCandidatesService {
     }
 
     @Override
-    public Pair<Float, Float> getRangeForStorage(List<NodeCandidate> nodeCandidates) {
-        Optional<Float> minValue = getHardwareMinValue(nodeCandidates, Hardware::getDisk);
-        Optional<Float> maxValue = getHardwareMaxValue(nodeCandidates, Hardware::getDisk);
-        return ImmutablePair.of(minValue.orElse(0f), maxValue.orElse(0f));
+    public Pair<Double, Double> getRangeForStorage(List<NodeCandidate> nodeCandidates) {
+        Optional<Double> minValue = getHardwareMinValue(nodeCandidates, Hardware::getDisk);
+        Optional<Double> maxValue = getHardwareMaxValue(nodeCandidates, Hardware::getDisk);
+        return ImmutablePair.of(minValue.orElse(0.0), maxValue.orElse(0.0));
     }
 
     @Override
@@ -66,7 +69,7 @@ public class NodeCandidateServiceImpl implements NodeCandidatesService {
     }
 
     @Override
-    public List<Float> getValuesForStorage(Map<Integer, List<NodeCandidate>> nodeCandidatesMap) {
+    public List<Double> getValuesForStorage(Map<Integer, List<NodeCandidate>> nodeCandidatesMap) {
         return getPossibleValues(nodeCandidatesMap, Hardware::getDisk);
     }
 
@@ -104,7 +107,7 @@ public class NodeCandidateServiceImpl implements NodeCandidatesService {
     }
 
     @Override
-    public List<Float> getValuesForStorage(List<NodeCandidate> nodeCandidates) {
+    public List<Double> getValuesForStorage(List<NodeCandidate> nodeCandidates) {
         return getPossibleValues(nodeCandidates, Hardware::getDisk);
     }
